Auth and middleware added

+++ routes/web.php
@@ -45,31 +45,31 @@ Route::get('/', [ListingController::class, 'index']);
 
 
 //Show Create Job form
-Route::get('/listings/create',[ListingController::class,'create']);
+Route::get('/listings/create',[ListingController::class,'create'])->middleware('auth');
 
 //Store Listing data
-Route::post('/listings',[ListingController::class,'store']);
+Route::post('/listings',[ListingController::class,'store'])->middleware('auth');
 
 //Show edit form
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
 //Update listing
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 
 //Delete listing
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
 
 //Show register form
-Route::get('/register',[UserController::class, 'create'] );
+Route::get('/register',[UserController::class, 'create'])->middleware('guest');
 
 //Create new user
 Route::post('/users',[UserController::class, 'store']);
 
 //Log user out
-Route::post('/logout',[UserController::class,'logout']);
+Route::post('/logout',[UserController::class,'logout'])->middleware('auth');
 
 //Show login form
-Route::get('/login', [UserController::class,'login']);
+Route::get('/login', [UserController::class,'login'])->name('login')->middleware('guest');//name renames the route to 'login' so that way the Authenticate.php knows who to reroute
 
 //Log in user
 Route::post('/users/authenticate',[UserController::class,'authenticate']);
